rearrange a few tests
