+++ tests/edgepy/test_glm.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 
 from inmoose.utils import rnbinom
-from inmoose.edgepy import addPriorCount, DGEList, glmFit, nbinomDeviance
+from inmoose.edgepy import DGEList, glmFit
 
 
 class test_DGEGLM(unittest.TestCase):
@@ -25,144 +25,6 @@ class test_DGEGLM(unittest.TestCase):
         self.assertIsNone(d.prior_count)
         self.assertIsNone(d.unshrunk_coefficients)
         self.assertIsNone(d.method)
-
-
-class test_nbinomDeviance(unittest.TestCase):
-    def setUp(self):
-        y = np.array(rnbinom(80, size=5, mu=20, seed=42)).reshape((20, 4))
-        self.y = np.vstack(([0, 0, 0, 0], [0, 0, 2, 2], y))
-
-    def test_nbinomDeviance(self):
-        dev_ref = np.array(
-            [
-                7.999999,
-                5.545177,
-                321.818901,
-                184.761910,
-                355.515748,
-                340.390858,
-                414.784086,
-                461.324010,
-                235.746631,
-                354.718673,
-                410.643778,
-                265.824737,
-                422.344282,
-                412.896876,
-                392.316278,
-                599.686176,
-                242.088733,
-                501.330767,
-                608.676351,
-                706.581453,
-                430.474599,
-                310.504051,
-            ]
-        )
-        dev = nbinomDeviance(self.y, np.ones(self.y.shape))
-        self.assertTrue(np.allclose(dev, dev_ref, atol=0, rtol=10 - 9))
-
-        dev_ref = np.array([8.454617])
-        dev = nbinomDeviance(np.array([1, 2, 3, 4]), np.ones(4), dispersion=np.zeros(4))
-        self.assertTrue(np.allclose(dev, dev_ref, atol=0, rtol=10 - 9))
-
-        with self.assertRaisesRegex(
-            ValueError, expected_regex="y is a matrix but mean is not"
-        ):
-            nbinomDeviance(self.y, np.ones((22,)))
-
-        with self.assertRaisesRegex(
-            ValueError, expected_regex="mean is a matrix but y is not"
-        ):
-            nbinomDeviance(np.array([1, 2, 3, 4]), np.ones((4, 3)))
-
-        with self.assertRaisesRegex(
-            ValueError, expected_regex="length of mean differs from that of y"
-        ):
-            nbinomDeviance(np.array([1, 2, 3, 4]), np.ones(3))
-
-
-class test_addPriorCount(unittest.TestCase):
-    def setUp(self):
-        y = np.array(rnbinom(80, size=5, mu=20, seed=42)).reshape((20, 4))
-        self.y = np.vstack(([0, 0, 0, 0], [0, 0, 2, 2], y))
-
-    def test_addPriorCount(self):
-        ref = np.array(
-            [
-                [0.9275688, 1.042111, 1.053341, 0.9769792],
-                [0.9275688, 1.042111, 3.053341, 2.9769792],
-                [25.9275688, 20.042111, 12.053341, 23.9769792],
-                [10.9275688, 18.042111, 11.053341, 17.9769792],
-                [14.9275688, 36.042111, 28.053341, 2.9769792],
-                [15.9275688, 8.042111, 29.053341, 29.9769792],
-                [20.9275688, 16.042111, 34.053341, 25.9769792],
-                [7.9275688, 36.042111, 29.053341, 28.9769792],
-                [15.9275688, 15.042111, 20.053341, 16.9769792],
-                [5.9275688, 25.042111, 23.053341, 30.9769792],
-                [35.9275688, 22.042111, 15.053341, 22.9769792],
-                [13.9275688, 23.042111, 16.053341, 19.9769792],
-                [34.9275688, 23.042111, 8.053341, 29.9769792],
-                [30.9275688, 33.042111, 18.053341, 13.9769792],
-                [26.9275688, 16.042111, 24.053341, 26.9769792],
-                [20.9275688, 26.042111, 39.053341, 38.9769792],
-                [17.9275688, 21.042111, 14.053341, 15.9769792],
-                [9.9275688, 49.042111, 36.053341, 8.9769792],
-                [41.9275688, 22.042111, 43.053341, 17.9769792],
-                [39.9275688, 34.042111, 41.053341, 25.9769792],
-                [21.9275688, 21.042111, 23.053341, 33.9769792],
-                [17.9275688, 20.042111, 24.053341, 18.9769792],
-            ]
-        )
-        oref = np.full((22, 4), [6.027929, 6.144366, 6.155085, 6.079828])
-        (yres, offsetres) = addPriorCount(self.y)
-        self.assertTrue(np.allclose(yres, ref, atol=0, rtol=1e-6))
-        self.assertTrue(np.allclose(offsetres, oref, atol=0, rtol=1e-6))
-
-        with self.assertRaisesRegex(
-            ValueError, expected_regex="count matrix must be numeric"
-        ):
-            addPriorCount(np.array(["foo", "bar"]))
-
-
-class test_predFC(unittest.TestCase):
-    def setUp(self):
-        y = np.array(rnbinom(80, size=5, mu=20, seed=42)).reshape((20, 4))
-        y = np.vstack(([0, 0, 0, 0], [0, 0, 2, 2], y))
-        self.group = np.array([1, 1, 2, 2])
-        self.d = DGEList(counts=y, group=self.group, lib_size=np.arange(1001, 1005))
-
-    def test_predFC(self):
-        design = np.array([[1, 0], [1, 0], [0, 1], [0, 2]])
-        with self.assertWarnsRegex(Warning, expected_regex="dispersion set to zero"):
-            res = self.d.predFC(design=design)
-        ref = np.array(
-            [
-                [-12.969746, -11.384905],
-                [-12.969746, -7.315955],
-                [-5.500709, -4.270495],
-                [-6.200135, -4.612279],
-                [-5.346262, -5.081762],
-                [-6.492589, -3.739418],
-                [-5.828765, -3.784606],
-                [-5.567436, -3.769349],
-                [-6.097954, -4.416721],
-                [-6.097954, -3.800063],
-                [-5.154532, -4.247723],
-                [-5.828765, -4.366038],
-                [-5.154532, -4.099175],
-                [-5.008310, -4.642870],
-                [-5.601993, -3.914313],
-                [-5.468468, -3.385668],
-                [-5.749148, -4.642870],
-                [-5.129109, -4.416721],
-                [-5.008310, -3.897303],
-                [-4.793386, -3.681757],
-                [-5.601993, -3.710232],
-                [-5.788407, -4.225364],
-            ]
-        )
-        self.assertTrue(np.allclose(res, ref, atol=1e-6, rtol=0))
 
 
 class test_glmFit(unittest.TestCase):
